Move nested messages in `ChatPreview` to the top level.
